ajout de tests dans fct, modif set

+++ include/mem.h
@@ -109,7 +109,7 @@ void print_byte_mem(mem memory, uint32_t adr, uint32_t val);
 int print_case_mem(mem memory,uint debut,uint fin);
 int find_val(mem memory, uint32_t adresse, uint32_t* res);
 int load_word(mem memory, uint32_t addresse, uint32_t wordtoload);
-int load_byte(mem memory, uint32_t adresse, uint32_t bytetoload);
+int load_byte(mem memory, uint32_t adresse, byte bytetoload);
 uint32_t swap_mot(uint32_t mot);
 #ifdef __cplusplus
 }
